Move asset saving from scheduler into own file.
