+++ src/scheduler.js
@@ -16,12 +16,8 @@
  */
 'use strict';
 
-const fs = require('fs');
-
-const log = require('./lib/log.js');
-const screenshotDump = require('./lib/screenshot-page.js');
+const assets = require('./lib/save-assets.js');
 const Gather = require('./gatherers/gather.js');
-const stringify = require('json-stringify-safe');
 
 function loadPage(driver, gatherers, options) {
   const loadPage = options.flags.loadPage;
@@ -84,38 +80,6 @@ function phaseRunner(gatherers) {
       return chain.then(_ => gatherFun(gatherer));
     }, Promise.resolve());
   };
-}
-
-// This isn't exposed functionality, and it's not core functionality, so skipped. */
-/* istanbul ignore next */
-function saveArtifacts(artifacts) {
-  const artifactsFilename = 'artifacts.log';
-  fs.writeFileSync(artifactsFilename, stringify(artifacts));
-  log.log('info', 'artifacts file saved to disk', artifactsFilename);
-}
-
-// This isn't exposed functionality, and it's not core functionality, so skipped. */
-/* istanbul ignore next */
-function getAssetFilename(assetName, url) {
-  const date = new Date();
-  const hostname = url.match(/^.*?\/\/(.*?)(:?\/|$)/)[1];
-  const filenamePrefix = hostname + '_' + date.toISOString();
-  return (filenamePrefix + assetName).replace(/[\/\?<>\\:\*\|":]/g, '-');
-}
-
-// This isn't exposed functionality, and it's not core functionality, so skipped. */
-/* istanbul ignore next */
-function saveAssets(options, artifacts) {
-  const url = options.url;
-  const traceFilename = getAssetFilename('.trace.json', url);
-
-  fs.writeFileSync(traceFilename, stringify(artifacts.traceContents, null, 2));
-  log.log('info', 'trace file saved to disk', traceFilename);
-
-  const screenshotsFilename = getAssetFilename('.screenshots.html', url);
-  const html = screenshotDump(screenshotsFilename, artifacts.screenshots);
-  fs.writeFileSync(screenshotsFilename, html);
-  log.log('info', 'screenshots saved to disk', screenshotsFilename);
 }
 
 function shouldRunPass(gatherers, phases) {
@@ -199,12 +163,12 @@ function run(gatherers, options) {
       // Ignoring these two flags since this functionality is not exposed by the module.
       /* istanbul ignore if */
       if (options.flags.saveArtifacts) {
-        saveArtifacts(artifacts);
+        assets.saveArtifacts(artifacts);
       }
 
       /* istanbul ignore if */
       if (options.flags.saveAssets) {
-        saveAssets(options, artifacts);
+        assets.saveAssets(options, artifacts);
       }
 
       return artifacts;
